+++ app/models/user.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import CheckConstraint, DateTime, String
+from sqlalchemy import CheckConstraint, DateTime, Identity, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.models.base import Base, PkType
@@ -10,7 +10,7 @@ class UserORM(Base):
     __tablename__ = "users"
     __table_args__ = (CheckConstraint("role IN ('admin', 'user')", name="role"),)
 
-    id: Mapped[int] = mapped_column(PkType, primary_key=True)
+    id: Mapped[int] = mapped_column(PkType, Identity(always=True), primary_key=True)
     email: Mapped[str] = mapped_column(String(255), nullable=False, unique=True)
     hashed_password: Mapped[str] = mapped_column(String(255), nullable=False)
     role: Mapped[str] = mapped_column(String(20), nullable=False, default="user")